requirements_linux.txt
anyio==3.7.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.98.0
filelock==3.9.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
pdf2image==1.16.3
Pillow==9.3.0
pydantic==1.10.9
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.11.1
torch==2.0.1+cpu
torchvision==0.15.2+cpu
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
